KDInputView: Preserve cursor when forcing case
If lowercase or uppercase is forced (`forceCase` option), the input
element's value is updated which will result in the user's current
cursor position to get reset to the end. This change preserves the
position by saving the position before updating the value, and then
setting it again to the old position after the value has been updated.

The setSelectionRange API is not supported in IE before 9, but a
polyfill is possible.